refactor: extract generic EventQueue into its own module
